refactor: reorganize cli and show modules (Issue #73)
