+++ pyproject.toml
@@ -112,9 +112,6 @@ addopts = [
   "-v",
   "--strict-markers",
   "--tb=short",
-  "--cov=src/pivot",
-  "--cov-report=term-missing",
-  "--cov-report=html",
 ]
 markers = [
   "slow: marks tests as slow (deselect with '-m \"not slow\"')",
