+++ src/pivot/show/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from pivot.show import metrics as metrics
+from pivot.show import params as params
+from pivot.show import plots as plots
